requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
requests==2.32.5
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
readability-lxml~=0.8.4.1
playwright~=1.55.0
dateparser~=1.2.2